remove duplicated popconfirm util logic
